both done , type and activite
