+++ server/routes/api.php
@@ -1,23 +1,21 @@
 <?php
 
-use App\Http\Controllers\AdministrateurController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\Api\ActiviteController;
+use App\Http\Controllers\Api\AdministrateurController;
 use App\Http\Controllers\Api\DemandeInscriptionController;
+use App\Http\Controllers\Api\DevisController;
 use App\Http\Controllers\Api\EnfantController;
-use App\Http\Controllers\DevisController;
+use App\Http\Controllers\Api\TypeActiviteController;
 
 Route::post('/login', [AuthController::class, 'login']);
 Route::post('/register', [AuthController::class, 'register']);
 
 
 
-Route::post('/devis/{id}/accept', [DevisController::class, 'acceptDevis']);
-Route::post('/devis/{id}/reject', [DevisController::class, 'rejectDevis']);
-Route::post('/admins/approve-demande/{id}', [AdministrateurController::class, 'approveDemande']);
-Route::post('/admin/reject-demande/{id}', [AdministrateurController::class, 'rejectDemande']);
+
 
 
 Route::group(['middleware' => 'auth:sanctum'], function () {
@@ -28,8 +26,11 @@ Route::group(['middleware' => 'auth:sanctum'], function () {
     Route::get('/users', [AuthController::class, 'index']);
     // for admins only and authenticated  
     //add middlewear check role 
+
     Route::apiResource('activites', ActiviteController::class);
-    Route::apiResource('type-activites', ActiviteController::class);
+    Route::apiResource('type-activites', TypeActiviteController::class);
+    Route::post('/admins/approve-demande/{id}', [AdministrateurController::class, 'approveDemande']);
+    Route::post('/admin/reject-demande/{id}', [AdministrateurController::class, 'rejectDemande']);
 
 
 
@@ -40,15 +41,10 @@ Route::group(['middleware' => 'auth:sanctum'], function () {
     //add middlewear check role 
     // for parents only and authenticated
     Route::apiResource('enfants', EnfantController::class);
-
-
-    
-    
-    
-    
     Route::apiResource('demande-Inscriptions', DemandeInscriptionController ::class); 
     Route::post('/devis/{id}/accept', [DevisController::class, 'acceptDevis']);
-    Route::post('/devis/{id}/reject', [App\Http\Controllers\DevisController::class, 'rejectDevis']);
+    Route::post('/devis/{id}/reject', [DevisController::class, 'rejectDevis']);
+
